refactor payrates
